-corrected indenting error -removed excess task_row.py file in components
